requirements.txt
streamlit==1.39.0
pandas==2.2.2
gspread==6.1.4
google-auth==2.35.0
gspread-dataframe==3.3.1
openpyxl==3.1.5
